# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SkillLevel {
  name: string;
  level: string;
  score: number;
  status: string;
}

export interface SkillGraphAnalysisInput {
  /** @minLength 1 */
  career: string;
  skills: SkillLevel[];
}

export interface RoadmapItem {
  week: number;
  topic: string;
  why: string;
  skills: string[];
  activity: string;
  effort: string;
}

export interface ProjectRecommendation {
  title: string;
  difficulty: string;
  description: string;
  skills: string[];
  why: string;
  stack: string[];
  outcome: string;
}

export interface SkillGraphAnalysis {
  readiness: number;
  strongSkills: string[];
  developingSkills: string[];
  priorityGaps: string[];
  whyGapsMatter: string;
  recommendedFocus: string;
  nextActions: string[];
  roadmap: RoadmapItem[];
  projects: ProjectRecommendation[];
}

export interface JobAnalysisInput {
  /** @minLength 1 */
  career: string;
  skills: SkillLevel[];
  /** @minLength 50 */
  jobDescription: string;
}

export interface JobAnalysis {
  role: string;
  requiredSkills: string[];
  preferredSkills: string[];
  technologies: string[];
  experienceExpectations: string;
  keywords: string[];
  covered: string[];
  partiallyCovered: string[];
  missing: string[];
  recommendedActions: string[];
}