# lib/api-zod/src/generated/types/jobAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobAnalysis {
  role: string;
  requiredSkills: string[];
  preferredSkills: string[];
  technologies: string[];
  experienceExpectations: string;
  keywords: string[];
  covered: string[];
  partiallyCovered: string[];
  missing: string[];
  recommendedActions: string[];
}