# lib/api-zod/src/generated/types/projectRecommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectRecommendation {
  title: string;
  difficulty: string;
  description: string;
  skills: string[];
  why: string;
  stack: string[];
  outcome: string;
}