# lib/api-zod/src/generated/types/roadmapItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoadmapItem {
  week: number;
  topic: string;
  why: string;
  skills: string[];
  activity: string;
  effort: string;
}