# lib/api-zod/src/generated/types/skillGraphAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectRecommendation } from './projectRecommendation';
import type { RoadmapItem } from './roadmapItem';

export interface SkillGraphAnalysis {
  readiness: number;
  strongSkills: string[];
  developingSkills: string[];
  priorityGaps: string[];
  whyGapsMatter: string;
  recommendedFocus: string;
  nextActions: string[];
  roadmap: RoadmapItem[];
  projects: ProjectRecommendation[];
}