# lib/api-zod/src/generated/types/skillGraphAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillLevel } from './skillLevel';

export interface SkillGraphAnalysisInput {
  /** @minLength 1 */
  career: string;
  skills: SkillLevel[];
}