# lib/api-zod/src/generated/types/skillLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillLevel {
  name: string;
  level: string;
  score: number;
  status: string;
}